工作流和个人空间申请中加入对应注释 修改个人空间申请样式

+++ src/views/data/expansion/index.vue
@@ -4,7 +4,8 @@
       <el-row>
         <el-form :inline="true"
                  :model="personalSpace"
-                 class="demo-form-inline">
+                 class="demo-form-inline"
+        >
           <el-form-item label="申请名称:">
             <el-input v-model="personalSpace.personalSpaceName"
                       placeholder="申请名称"></el-input>
@@ -64,13 +65,15 @@
                 border
                 style="width: 100%"
                 height="calc(100vh - 290px)"
-                @selection-change="handleSelectionChange">
+                @selection-change="handleSelectionChange"
+      >
         <el-table-column type="selection"
                          width="55">
         </el-table-column>
         <el-table-column fixed
                          prop="personalSpaceName"
-                         label="申请名称">
+                         label="申请名称"
+                         show-overflow-tooltip>
         </el-table-column>
         <el-table-column prop="personalSpaceCapacity"
                          label="扩容大小">
@@ -118,13 +121,15 @@
                :visible.sync="openInsertDialog"
                width="60%">
       <el-form :model="personalSpace"
-               class="demo-form-inline"
-               label-width="80px">
-        <el-form-item label="申请名称">
+               class="demo-ruleForm"
+               label-width="80px"
+               :rules="rules"
+      >
+        <el-form-item label="申请名称" class="item-b" prop="personalSpaceName">
           <el-input v-model="personalSpace.personalSpaceName"
                     placeholder="申请名称"></el-input>
         </el-form-item>
-        <el-form-item label="扩容容量">
+        <el-form-item label="扩容容量" class="item-b" prop="personalSpaceCapacity">
           <div style="display: flex">
             <el-input v-model="personalSpace.personalSpaceCapacity"
                       placeholder="扩容容量"
@@ -132,9 +137,10 @@
                       :max="1024"
                       :min="0"
                       @input="inputChange"
-                      style="width: 540px"></el-input>
+                      style="width: 80%"></el-input>
             <el-select v-model="personalSpaceCapacityNeed"
-                       placeholder="容量单位" style="margin-left: 10px">
+                       placeholder="容量单位" style="margin-left: 10px"
+            >
               <el-option label="GB"
                          value="GB"></el-option>
               <el-option label="MB"
@@ -150,14 +156,15 @@
 <!--                    style="width: 670px"></el-input>-->
 <!--          <el-button type="primary">选择</el-button>-->
 <!--        </el-form-item>-->
-        <el-row type="flex"
-                justify="end">
-          <el-button type="primary"
-                     @click="onSubmit">保存</el-button>
-          <el-button type="primary"
-                     @click="closeInsertDialog">关闭</el-button>
-        </el-row>
+
       </el-form>
+      <el-row type="flex"
+              justify="end">
+        <el-button type="primary"
+                   @click="onSubmit">保存</el-button>
+        <el-button type="primary"
+                   @click="closeInsertDialog">关闭</el-button>
+      </el-row>
     </el-dialog>
 
     <!--    修改弹窗-->
@@ -293,13 +300,17 @@
     data(){
       return{
         personalSpace: {
-          personalSpaceName: '',
-          personalSpaceApplication: '',
-          personalSpaceType:'',
-          personalSpaceDate:'',
-          personalSpaceCapacity:'',
-          personalSpaceStatus:'',
+          personalSpaceName: '',//个人空间申请名称
+          personalSpaceApplication: '',//申请人
+          personalSpaceType:'',//业务所处状态
+          personalSpaceDate:'',//申请日期
+          personalSpaceCapacity:'',//申请的容量
+          // personalSpaceStatus:'',//根据原型画过来的 暂时没用到捏
           // personalSpaceApproving:'',
+        },
+        rules: {
+          personalSpaceName: [{ required: true, message: "请输入", trigger: "blur" }],
+          personalSpaceCapacity: [{ required: true, message: "请输入", trigger: "blur" }],
         },
         personalSpaceDataList: [],//查询并展示的数据集合
         openInsertDialog: false,//打开添加弹窗
@@ -322,13 +333,13 @@
         },
         flowItem: {
           //动态赋值
-          wftype: "auditNotice",
-          applyUuid: "",
+          wftype: "auditNotice",//当前业务所属的流程的id
+          applyUuid: "",//申请单主键id
           detailUuids: "",
-          applyTitle: "",
-          workEffortId: "",
-          appDataUuid: "",
-          versionUuid: "",
+          applyTitle: "",//请求的标题
+          workEffortId: "",//节点id
+          appDataUuid: "",//业务主键id
+          versionUuid: "",//版本id
           isSecond: false,
           temp1: "",
         },
@@ -407,17 +418,20 @@
       backToUpPage(){
         this.$router.go(-1)
       },
+      //打开新增弹窗
       joinInsertDialog(){
         this.$nextTick(()=>{
           this.clearParams()
         })
         this.openInsertDialog = true
       },
+      //关闭新增弹窗
       closeInsertDialog(){
         this.openInsertDialog = false
         this.openUpdateDialog = false
         this.clearParams()
       },
+      //新增页面保存方法
       onSubmit(){
         const capacity1 = this.personalSpace.personalSpaceCapacity;
         const capacity2 = this.personalSpaceCapacityNeed
@@ -452,45 +466,49 @@
             this.$notify.error("添加失败")
           })
       },
+      // 删除按钮
       onDelete(){
         var result = ''
-        // this.personalSpaceSelectionList.forEach((value,index)=>{
-        //   if(value.personalSpaceType != '草稿'){
-        //     this.$notify.warning("只有草稿状态才可以删除")
-        //     result = 'break'
-        //   }
-        // })
-        // if(result == 'break'){
-        //   return
-        // }
+        this.personalSpaceSelectionList.forEach((value,index)=>{
+          if(value.personalSpaceType != '草稿'){
+            this.$notify.warning("只有草稿状态才可以删除")
+            result = 'break'
+          }
+        })
+        if(result == 'break'){
+          return
+        }
         const params = this.personalSpaceUuidList
-        // if(params.length == 0){
-        //   this.$message({
-        //     type: 'warning',
-        //     message: '未选择删除对象'
-        //   })
-        // }else{
-        //   this.$confirm("是否删除该条数据?", "提示", {
-        //     confirmButtonText: "确定",
-        //     cancelButtonText: "取消",
-        //     type: "warning",
-        //     center: true,
-        //   }).then(() => {
+        if(params.length == 0){
+          this.$message({
+            type: 'warning',
+            message: '未选择删除对象'
+          })
+        }else{
+          this.$confirm("是否删除该条数据?", "提示", {
+            confirmButtonText: "确定",
+            cancelButtonText: "取消",
+            type: "warning",
+            center: true,
+          }).then(() => {
             deletePersonalSpace(params).then((res)=>{
               this.$notify.success("删除成功")
               this.initPersonalSpaceData()
             })
-          // })
-        // }
+          })
+        }
       },
+      //分页组件的页大小change事件
       handleSizeChange(val){
         this.query.pageSize = val
         this.initPersonalSpaceData()
       },
+      //分页组件当前页change事件
       handleCurrentChange(val){
         this.query.pageNo = val
         this.initPersonalSpaceData()
       },
+      //清空数据
       clearParams(){
         this.personalSpace = {
           personalSpaceName: '',
@@ -498,11 +516,12 @@
           personalSpaceType:'',
           personalSpaceDate:'',
           personalSpaceCapacity:'',
-          personalSpaceStatus:'',
+          // personalSpaceStatus:'',
           // personalSpaceApproving:'',
         }
         this.personalSpaceCapacityNeed = ''
       },
+      //容量大小输入限制
       inputChange(val){
         if(val<0){
           this.personalSpace.personalSpaceCapacity = 0
@@ -511,6 +530,7 @@
           this.personalSpace.personalSpaceCapacity = 1024
         }
       },
+      //提交按钮
       handleEvent(){
         if(this.personalSpaceSelectionList.length >1){
           this.$notify.warning("一次只能提交一条数据")
@@ -529,6 +549,7 @@
         this.flowItem.applyTitle=this.temp.personalSpaceName;
         this.dialogFlowItemShow=true;
       },
+      // 打开修改弹窗
       joinUpdateDialog(){
         const param = this.personalSpaceUuidList[0]
         if(param == undefined || param == null || param.length == 0){
@@ -536,7 +557,7 @@
           return
         }
         if(this.personalSpaceSelectionList.length >1){
-          this.$notify.warning("一次只能修改一个对象")
+          this.$notify.warning("一次只能修改一条数据")
           return
         }
         if(this.personalSpaceSelectionList[0].personalSpaceType !== '草稿'){
@@ -548,6 +569,7 @@
         this.$refs.expansionUpdatePage.queryByUuid(param)
 
       },
+      //修改方法
       onUpdate(personalSpace,personalSpaceCapacityNeed){
         const capacity1 = personalSpace.personalSpaceCapacity
         const capacity2 = personalSpaceCapacityNeed
@@ -556,9 +578,6 @@
         updatePersonalSpace(param)
         .then((res)=>{
           this.$notify.success("修改成功")
-          // setTimeout(() => {
-          //     location.reload()
-          // }, 20);
           this.openUpdateDialog = false
           this.initPersonalSpaceData()
 
@@ -569,6 +588,7 @@
         })
 
       },
+      //打开流程弹窗
       handle(){
         this.dialogFlowItemShow=true;
       },
@@ -592,6 +612,7 @@
           })
         }, 20);
       },
+      //打开流程跟踪弹窗
       todoOpinionList(row){
         //正常情况下 applyUuid 为 审批单主键
         //在业务页面 因为拿到 审批单主键比较麻烦
@@ -604,12 +625,25 @@
   }
 </script>
 
-<style scoped>
+<style lang="scss" scoped>
 .padding10 {
   padding: 10px;
   box-sizing: border-box;
 }
 .header_Filter {
   width: 100%;
+}
+.demo-ruleForm {
+  margin-top:20px;
+  display: flex;
+  flex-wrap: wrap;
+  ::v-deep .el-form-item {
+    margin-bottom: 22px !important;
+  }
+  ::v-deep .el-form-item__label {
+    text-align: right;
+    vertical-align: middle;
+    float: left !important;
+  }
 }
 </style>
